+++ CHANGELOG.md
@@ -52,7 +52,7 @@
 
 ## stream-chat-android-ui-common
 ### 🐞 Fixed
-
+- Using the same `MessageListItemViewHolderFactory` of `MessageListView` in `MessageOptionsDialogFragment`
 ### ⬆️ Improved
 
 ### ✅ Added